+++ theory/src/main/java/at/ainf/theory/storage/Partition.java
@@ -17,7 +17,7 @@ public class Partition<T> {
     public Set<AxiomSet<T>> dz = new LinkedHashSet<AxiomSet<T>>();
 
     public Set<T> partition;
-    public double score = Double.MAX_VALUE;
+    public BigDecimal score = BigDecimal.valueOf(Double.MAX_VALUE);
     public BigDecimal difference = new BigDecimal(Double.MAX_VALUE);
     public boolean isVerified = false;
